Refactored some some code, added documentation, some renaming and also set the txt file for requirements.
I changed the false naming of faulty and healthy to IRF and BF. I also added a txt file but I still didn't put the actual modules yet. Changed the file name of the tutorial to a shorted more descriptive name. I also changed some stuff in the readme file for future updates.